authentication service/requirements.txt
asgiref==3.4.1
Django==3.2.9
pip==21.3.1
pytz==2021.3
setuptools==59.5.0
sqlparse==0.4.2
wheel==0.37.0
djangorestframework==3.12.4
markdown==3.3.6
django-filter==21.1
djangorestframework-simplejwt==5.0.0
